Move enum types to enum_types.py
