+++ dadguide/dadguide.py
@@ -27,7 +27,7 @@ from .database_loader import load_database
 from .idtest_mixin import IdTest
 from .models.monster_model import MonsterModel
 from .models.monster_stats import monster_stats, MonsterStatModifierInput
-from .common.enums import DEFAULT_SERVER, SERVERS
+from .models.enum_types import DEFAULT_SERVER, SERVERS
 from .monster_index import MonsterIndex
 
 logger = logging.getLogger('red.padbot-cogs.dadguide')
