+++ dadguide/find_monster_tokens.py
@@ -2,7 +2,7 @@ import re
 
 from dadguide.models.monster_model import MonsterModel
 from dadguide.token_mappings import AWAKENING_TOKENS, AWOKEN_SKILL_MAP, PLUS_AWOKENSKILL_MAP
-from dadguide.common.enums import AwokenSkills
+from dadguide.models.enum_types import AwokenSkills
 
 
 def regexlist(tokens):
