+++ dadguide/idtest_mixin.py
@@ -7,7 +7,7 @@ from tsutils.enums import Server
 from redbot.core import commands, Config, checks
 from redbot.core.utils.chat_formatting import box, pagify
 
-from dadguide.common.enums import DEFAULT_SERVER
+from dadguide.models.enum_types import DEFAULT_SERVER
 from dadguide.monster_index import MonsterIndex
